# api/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "description": "Cryptographically random, named, expiring, revocable role-scoped credential",
        "scheme": "bearer",
        "type": "http"
      },
      "browserSession": {
        "description": "Short-lived HttpOnly SameSite=Strict session with Origin and CSRF checks on mutations",
        "in": "cookie",
        "name": "bridge_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "No public management ingress. Explicit local owner bootstrap/recovery has no TCP endpoint. Plans expire after ten minutes; source and external effects are separate. GPU qualification remains independently enforced by the host executor.",
    "title": "Spry.ai Workstation Bridge private management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/v1/audit": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "action": {
                        "type": "string"
                      },
                      "actor": {
                        "type": "string"
                      },
                      "at": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "object": {
                        "type": "string"
                      },
                      "result": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "at",
                      "actor",
                      "action",
                      "object",
                      "result"
                    ],
                    "type": "object"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "summary": "Read bounded operation and identity audit records",
        "x-minimum-role": "owner"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "parameters": [
          {
            "description": "Required with browser session cookies; Origin must match configured management origin. Bearer requests do not require browser headers.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "credential": {
                    "type": "string"
                  }
                },
                "required": [
                  "credential"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "actor": {
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "role"
                      ],
                      "type": "object"
                    },
                    "csrf": {
                      "type": "string"
                    },
                    "expires_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "actor",
                    "csrf",
                    "expires_at",
                    "mode"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "security": [],
        "summary": "Exchange a local scoped credential for a browser session",
        "x-minimum-role": "credential"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "parameters": [
          {
            "description": "Required with browser session cookies; Origin must match configured management origin. Bearer requests do not require browser headers.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {},
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": [
                    "object",
                    "null"
                  ]
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "summary": "Revoke browser session",
        "x-minimum-role": "viewer"
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "actor": {
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "role"
                      ],
                      "type": "object"
                    },
                    "csrf": {
                      "type": "string"
                    },
                    "expires_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "actor",
                    "csrf",
                    "expires_at",
                    "mode"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "summary": "Inspect authenticated browser identity",
        "x-minimum-role": "viewer"
      }
    },
    "/api/v1/builds": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "output": {
                        "type": "string"
                      },
                      "qualification": {
                        "type": "string"
                      },
                      "reason": {
                        "type": "string"
                      },
                      "revision": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "revision",
                      "status",
                      "reason",
                      "output",
                      "qualification"
                    ],
                    "type": "object"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "summary": "List reviewed build recipes and explicit refusals",
        "x-minimum-role": "viewer"
      }
    },
    "/api/v1/caches": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "budget_bytes": {
                        "format": "int64",
                        "type": "integer"
                      },
                      "cleanup_preview": {
                        "items": {
                          "type": "string"
                        },
                        "type": [
                          "array",
                          "null"
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "used_bytes": {
                        "format": "int64",
                        "type": "integer"
                      }
                    },
                    "required": [
                      "name",
                      "used_bytes",
                      "budget_bytes",
                      "status",
                      "cleanup_preview"
                    ],
                    "type": "object"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "Successful management response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Structured refusal or failure; 401 authentication, 403 authorization/CSRF, 409 drift/conflict, 413 body bound, 429 capacity/rate, 503 unavailable"
          }
        },
        "summary": "List managed cache usage and cleanup preview",
        "x-minimum-role": "viewer"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "browserSession": []
    }
  ]
}